updated posts

+++ app/routes/posts.py
@@ -1,9 +1,12 @@
+from typing import Optional
+
 from fastapi import APIRouter, HTTPException, status
 
 from fastapi import FastAPI, Depends
 from sqlalchemy.orm import Session
 from database import create_tables, get_db
 import models, schemas
+from app.utils import oauth2
 
 
 
@@ -15,8 +18,8 @@ router = APIRouter(
 )
 
 @router.get("/" , response_model=list[schemas.Post])
-async def get_posts(db : Session = Depends(get_db)):
-    return db.query(models.Post).all()
+async def get_posts(db : Session = Depends(get_db), limit: int = 10, skip: int = 0, search: str = ""):
+    return db.query(models.Post).filter(models.Post.title.contains(search)).offset(skip).limit(limit).all()
 
 @router.get("/{id}" , response_model=schemas.Post)
 async def get_post(id: int, db : Session = Depends(get_db)):
@@ -26,26 +29,34 @@ async def get_post(id: int, db : Session = Depends(get_db)):
     return db_post
 
 @router.post("/", response_model=schemas.Post, status_code=status.HTTP_201_CREATED)
-async def create_posts(post: schemas.PostCreate, db : Session = Depends(get_db)):
+async def create_posts(post: schemas.PostCreate, db : Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
     db_post = models.Post(**post.model_dump())
+    db_post.user_id = current_user.id
     db.add(db_post)
     db.commit()
     db.refresh(db_post)
     return db_post
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(id: int, db : Session = Depends(get_db)):
+async def delete_post(id: int, db : Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
     db_post =  db.query(models.Post).filter(models.Post.id == id).first()
     if not db_post:
         raise HTTPException(status_code=404, detail=f"Post with id {id} not found")
+    
+    if db_post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this post")
     db.delete(db_post)
     db.commit()
     return
+
+
 @router.put("/{id}" , response_model=schemas.Post)
-async def update_post(id: int, post: schemas.PostCreate, db : Session = Depends(get_db)):
+async def update_post(id: int, post: schemas.PostCreate, db : Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
     db_post =  db.query(models.Post).filter(models.Post.id == id).first()
     if not db_post:
         raise HTTPException(status_code=404, detail=f"Post with id {id} not found")
+    if db_post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
     for key, value in post.model_dump().items():
         setattr(db_post, key, value)
     db.commit()
